flowproc-project/requirements-dev.txt
# Development dependencies for FlowProcessor
# Install with: pip install -r requirements-dev.txt

# Core dependencies (from requirements.txt)
-r requirements.txt

# Development tools
pytest>=8.0.0
pytest-qt>=4.3.0
pytest-cov>=6.0.0
pytest-mock>=3.13.0
pytest-qt-plugins>=3.3.0
pytest-qt-plugins-pyside6>=3.3.0
pytest-qt-plugins-pyside6-essentials>=3.3.0
pytest-qt-plugins-pyside6-addons>=3.3.0

# Code quality tools
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=2.0.0

# Type checking
types-PyYAML>=6.0.0
types-python-dateutil>=2.8.0

# Performance profiling
memory-profiler>=0.61.0
line-profiler>=4.1.0 